refactor : id, createAt, updatedAt 공통 컬럼 관리 엔티티 적용
