+++ src/main/java/qna/domain/User.java
@@ -3,20 +3,12 @@ package qna.domain;
 import qna.UnAuthorizedException;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Entity
 @Table(name = "user")
-public class User {
+public class User extends BaseEntity {
     public static final GuestUser GUEST_USER = new GuestUser();
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
 
     @Column(name = "email", length = 50)
     private String email;
@@ -26,9 +18,6 @@ public class User {
 
     @Column(name = "password", length = 20, nullable = false)
     private String password;
-
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
 
     @Column(name = "user_id", length = 20, nullable = false, unique = true)
     private String userId;
@@ -41,13 +30,12 @@ public class User {
     }
 
     public User(Long id, String userId, String password, String name, String email) {
+        super(id);
         validate(userId, password, name);
-        this.id = id;
         this.userId = userId;
         this.password = password;
         this.name = name;
         this.email = email;
-        this.createdAt = LocalDateTime.now();
     }
 
     private void validate(String userId, String password, String name) {
@@ -96,10 +84,6 @@ public class User {
         return false;
     }
 
-    public Long getId() {
-        return id;
-    }
-
     public String getUserId() {
         return userId;
     }
@@ -114,27 +98,6 @@ public class User {
 
     public String getEmail() {
         return email;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "id=" + id +
-                ", createdAt=" + createdAt +
-                ", email='" + email + '\'' +
-                ", name='" + name + '\'' +
-                ", password='" + password + '\'' +
-                ", updatedAt=" + updatedAt +
-                ", userId='" + userId + '\'' +
-                '}';
     }
 
     private static class GuestUser extends User {
